+++ ElvUI_MerathilisUI/Core/Options/Misc.lua
@@ -148,10 +148,10 @@ options.spellAlert = {
 			type = "toggle",
 			name = L["Enable"]
 		},
-		visability = {
+		visibility = {
 			order = 4,
 			type = "toggle",
-			name = L["Visablity"],
+			name = L["Visibility"],
 			desc = L["Enable/Disable the spell activation alert frame."],
 			get = function(info)
 				return C_CVar_GetCVarBool("displaySpellActivationOverlays")