+++ ElvUI_MerathilisUI/Core/Settings/Global.lua
@@ -1,11 +1,11 @@
 local MER, F, E, L, V, P, G = unpack(select(2, ...))
 
 G.core = {
-	loginMsg = true,
-	logLevel = 2,
 	compatibilityCheck = true,
+	cvarAlert = false,
+	logLevel = 2,
+	loginMsg = true,
 	fixLFG = true,
-	fixCVAR = true,
 }
 
 G.mail = {