+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ### Changes ###
 
-  * Fixed an issue with Instance Indicator on the Minimap
-  * Fixed RaiderIO Skin
-  * Small Updates to the UIWidgets
+  * Use a Global Variable on AlreadyKnown
+  * Fixed an error on the Encounter Journal
+  * Added new Phial's to the RBReminder
+  * Changed the CVar Alert to default off
+  * Updated Event Tracker
+  * Removed some font settings on blizzards widgets